+++ graph/telemetry/istio/appender/service_entry.go
@@ -23,7 +23,7 @@ func (a ServiceEntryAppender) Name() string {
 }
 
 // AppendGraph implements Appender
-func (a ServiceEntryAppender) AppendGraph(trafficMap graph.TrafficMap, globalInfo *GlobalInfo, namespaceInfo *NamespaceInfo) {
+func (a ServiceEntryAppender) AppendGraph(trafficMap graph.TrafficMap, globalInfo *graph.AppenderGlobalInfo, namespaceInfo *graph.AppenderNamespaceInfo) {
 	if len(trafficMap) == 0 {
 		return
 	}
@@ -31,7 +31,7 @@ func (a ServiceEntryAppender) AppendGraph(trafficMap graph.TrafficMap, globalInf
 	a.applyServiceEntries(trafficMap, globalInfo, namespaceInfo)
 }
 
-func (a ServiceEntryAppender) applyServiceEntries(trafficMap graph.TrafficMap, globalInfo *GlobalInfo, namespaceInfo *NamespaceInfo) {
+func (a ServiceEntryAppender) applyServiceEntries(trafficMap graph.TrafficMap, globalInfo *graph.AppenderGlobalInfo, namespaceInfo *graph.AppenderNamespaceInfo) {
 	for _, n := range trafficMap {
 		// only a service node can be a service entry
 		if n.NodeType != graph.NodeTypeService {
@@ -55,9 +55,9 @@ func (a ServiceEntryAppender) applyServiceEntries(trafficMap graph.TrafficMap, g
 // across all accessible namespaces in the cluster. All ServiceEntries are needed because
 // Istio does not distinguish where a ServiceEntry is created when routing traffic (i.e.
 // a ServiceEntry can be in any namespace and it will still work).
-func (a ServiceEntryAppender) getServiceEntry(service string, globalInfo *GlobalInfo) (string, bool) {
-	if globalInfo.ServiceEntries == nil {
-		globalInfo.ServiceEntries = make(map[string]string)
+func (a ServiceEntryAppender) getServiceEntry(service string, globalInfo *graph.AppenderGlobalInfo) (string, bool) {
+	if globalInfo.Telemetry["ServiceEntries"] == nil {
+		globalInfo.Telemetry["ServiceEntries"] = make(map[string]string)
 
 		for ns := range a.AccessibleNamespaces {
 			istioCfg, err := globalInfo.Business.IstioConfig.GetIstioConfigList(business.IstioConfigCriteria{
@@ -73,14 +73,14 @@ func (a ServiceEntryAppender) getServiceEntry(service string, globalInfo *Global
 						location = "MESH_INTERNAL"
 					}
 					for _, host := range entry.Spec.Hosts.([]interface{}) {
-						globalInfo.ServiceEntries[host.(string)] = location
+						globalInfo.Telemetry["ServiceEntries"].(map[string]string)[host.(string)] = location
 					}
 				}
 			}
 		}
-		log.Tracef("Found [%v] service entries", len(globalInfo.ServiceEntries))
+		log.Tracef("Found [%v] service entries", len(globalInfo.Telemetry["ServiceEntries"].(map[string]string)))
 	}
 
-	location, ok := globalInfo.ServiceEntries[service]
+	location, ok := globalInfo.Telemetry["ServiceEntries"].(map[string]string)[service]
 	return location, ok
 }
