+++ graph/types.go
