+++ hack/istio/bookinfo-details-delay.yaml
@@ -1,0 +1,32 @@
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: details-delay-abort-vs
+spec:
+  hosts:
+  - ratings
+  http:
+  - fault:
+      delay:
+        fixedDelay: 7s
+        percent: 50
+    route:
+    - destination:
+        host: details
+        subset: details-delay-abort-dr-subset-v1
+  - route:
+    - destination:
+        host: details
+        subset: details-delay-abort-dr-subset-v1
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: DestinationRule
+metadata:
+  name: details-delay-abort-dr
+spec:
+  host: details
+  subsets:
+  - name: details-delay-abort-dr-subset-v1
+    labels:
+      version: v1
