+++ hack/istio/bookinfo-productpage-v1-cb.yaml
@@ -1,0 +1,22 @@
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: DestinationRule
+metadata:
+  namespace: bookinfo
+  name: circuitbreaker-pp-v1
+spec:
+  host: productpage
+  subsets:
+  - name: circuitbreaker-pp-v1
+    labels:
+      version: v1
+    trafficPolicy:
+      connectionPool:
+        http:
+          http1MaxPendingRequests: 1
+          maxRequestsPerConnection: 1
+      outlierDetection:
+        consecutiveErrors: 1
+        interval: 1s
+        baseEjectionTime: 10s
+        maxEjectionPercent: 50
