+++ hack/istio/bookinfo-ratings-abort.yaml
@@ -1,0 +1,32 @@
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: ratings-abort-vs
+spec:
+  hosts:
+  - ratings
+  http:
+  - fault:
+      abort:
+        httpStatus: 555
+        percent: 50
+    route:
+    - destination:
+        host: ratings
+        subset: ratings-abort-dr-subset-v1
+  - route:
+    - destination:
+        host: ratings
+        subset: ratings-abort-dr-subset-v1
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: DestinationRule
+metadata:
+  name: ratings-abort-dr
+spec:
+  host: ratings
+  subsets:
+  - name: ratings-abort-dr-subset-v1
+    labels:
+      version: v1
