+++ hack/istio/bookinfo-rename/details-abort.yaml
@@ -1,0 +1,34 @@
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: details-abort-vs
+  namespace: bookinfo2
+spec:
+  hosts:
+  - details
+  http:
+  - fault:
+      abort:
+        httpStatus: 555
+        percent: 100
+    route:
+    - destination:
+        host: details
+        subset: details-abort-dr-subset-v1
+  - route:
+    - destination:
+        host: details
+        subset: details-abort-dr-subset-v1
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: DestinationRule
+metadata:
+  name: details-abort-dr
+  namespace: bookinfo2
+spec:
+  host: details
+  subsets:
+  - name: details-abort-dr-subset-v1
+    labels:
+      version: v1
