+++ hack/istio/bookinfo-reviews-other-ns.yaml
@@ -1,0 +1,45 @@
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  namespace: bookinfo
+  name: reviews-other-ns-vs
+spec:
+  hosts:
+  - reviews
+  http:
+  - route:
+    - destination:
+        host: reviews
+        subset: reviews-v1
+      weight: 25
+    - destination:
+        host: reviews
+        subset: reviews-v2
+      weight: 25
+    - destination:
+        host: reviews
+        subset: reviews-v3
+      weight: 25
+    - destination:
+        host: reviews.kiali-test-reviews.svc.cluster.local
+      weight: 25
+
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: DestinationRule
+metadata:
+  namespace: bookinfo
+  name: reviews-bookinfo-dr
+spec:
+  host: reviews.bookinfo.svc.cluster.local
+  subsets:
+  - name: reviews-v1
+    labels:
+      version: v1
+  - name: reviews-v2
+    labels:
+      version: v2
+  - name: reviews-v3
+    labels:
+      version: v3
+
