+++ hack/istio/istio-ingress-vs.yaml
@@ -1,0 +1,16 @@
+---
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: ingress-vs
+spec:
+  hosts:
+  - istio-ingressgateway.istio-system.svc
+  http:
+  - match:
+      - uri:
+          prefix: /
+    route:
+    - destination:
+        host: istio-ingressgateway
+
